+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Upgrade resources to `v0488`
 - Add mission APIs
 - Farm Hall of Fame reward on Saturday
+- Disable `backtrace`
 
 ## `3.0b2`